requirements.txt
amqp==5.2.0
anyio==4.6.2.post1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.1.2
django-allauth==65.2.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-debug-toolbar==4.4.6
django-timezone-field==7.0
django-user-agents==0.4.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
kombu==5.4.2
pillow==11.0.0
prompt_toolkit==3.0.48
pyTelegramBotAPI==4.24.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.6
qrcode==8.0
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
ua-parser==0.18.0
urllib3==2.2.3
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13

whitenoise==6.6.0
gunicorn==21.2.0 